requirements.txt
fastapi
uvicorn
pillow
pytesseract
